This is the 3.2.56 stable release

Conflicts:
	drivers/md/raid5.c

+++ kernel/time/timekeeping.c
@@ -1173,7 +1173,7 @@ void get_monotonic_boottime(struct timespec *ts)
 	} while (read_seqcount_retry(&xtime_seq, seq));
 
 	set_normalized_timespec(ts, ts->tv_sec + tomono.tv_sec + sleep.tv_sec,
-			ts->tv_nsec + tomono.tv_nsec + sleep.tv_nsec + nsecs);
+		(s64)ts->tv_nsec + tomono.tv_nsec + sleep.tv_nsec + nsecs);
 }
 EXPORT_SYMBOL_GPL(get_monotonic_boottime);
